+++ kibot/out_hpgl_sch_print.py
@@ -41,7 +41,8 @@ class HPGL_SCH_PrintOptions(Any_SCH_PrintOptions):
 class HPGL_SCH_Print(BaseOutput):  # noqa: F821
     """ HPGL Schematic Print (Hewlett & Packard Graphics Language)
         Exports the schematic to the most common plotter format.
-        This output is what you get from the 'File/Plot' menu in eeschema. """
+        This output is what you get from the 'File/Plot' menu in eeschema.
+        Supports the image replacement using the prefix indicated by the `sch_image_prefix` global variable """
     __doc__ += FONT_HELP_TEXT
 
     def __init__(self):