+++ kibot/out_svg_sch_print.py
@@ -35,7 +35,8 @@ class SVG_SCH_Print(BaseOutput):  # noqa: F821
     """ SVG Schematic Print
         Exports the schematic in a vectorized graphics format.
         This is a format to document your schematic.
-        This output is what you get from the 'File/Plot' menu in eeschema. """
+        This output is what you get from the 'File/Plot' menu in eeschema.
+        Supports the image replacement using the prefix indicated by the `sch_image_prefix` global variable """
     __doc__ += FONT_HELP_TEXT
 
     def __init__(self):